one server to handle storage and viewing
One server should handle all of the requests to store and view the results.

* merge report server api and implementation into the report viewer
* only one common server is started to store and view the results
* authentication is introduced for the storage api too
* store and check should use a codechecker server host and port
  no direct connection to the database is available
* database arguments in store and check are deprecated
* authentication and client api was factored out from the command
  line client modul to a common client module
* new test which tries to store results to an authenticated server
