+++ tests/functional/delete_runs/__init__.py
@@ -86,6 +86,10 @@ def setup_package():
     # Extend the checker configuration with the port numbers.
     codechecker_cfg.update(host_port_cfg)
 
+    # Start the CodeChecker server.
+    print("Starting server to get results")
+    _start_server(codechecker_cfg, test_config, False)
+
     for i in range(0, 5):
         # Clean the test project, if needed by the tests.
         ret = project.clean(test_project)
@@ -99,10 +103,6 @@ def setup_package():
         if ret:
             sys.exit(1)
 
-        if pg_db_config:
-            print("Waiting for PostgreSQL to stop.")
-            codechecker.wait_for_postgres_shutdown(TEST_WORKSPACE)
-
         print("Analyzing the test project was successful {}.".format(str(i)))
 
     # Save the run names in the configuration.
@@ -113,10 +113,6 @@ def setup_package():
 
     # Export the test configuration to the workspace.
     env.export_test_cfg(TEST_WORKSPACE, test_config)
-
-    # Start the CodeChecker server.
-    print("Starting server to get results")
-    _start_server(codechecker_cfg, test_config, False)
 
 
 def teardown_package():
