+++ tests/functional/suppress/__init__.py
@@ -71,6 +71,10 @@ def setup_package():
     if ret:
         sys.exit(ret)
 
+    # Start the CodeChecker server.
+    print("Starting server to get results")
+    _start_server(codechecker_cfg, test_config, False)
+
     test_project_name = project_info['name'] + '_' + uuid.uuid4().hex
 
     ret = codechecker.check(codechecker_cfg,
@@ -81,19 +85,11 @@ def setup_package():
         sys.exit(1)
     print("Analyzing the test project was successful.")
 
-    if pg_db_config:
-        print("Waiting for PostgreSQL to stop.")
-        codechecker.wait_for_postgres_shutdown(TEST_WORKSPACE)
-
     codechecker_cfg['run_names'] = [test_project_name]
 
     test_config['codechecker_cfg'] = codechecker_cfg
 
     env.export_test_cfg(TEST_WORKSPACE, test_config)
-
-    # Start the CodeChecker server.
-    print("Starting server to get results")
-    _start_server(codechecker_cfg, test_config, False)
 
 
 def teardown_package():
