+++ tests/functional/update/__init__.py
@@ -70,6 +70,10 @@ def setup_package():
 
     codechecker_cfg.update(host_port_cfg)
 
+    # Start the CodeChecker server.
+    print("Starting server to get results")
+    _start_server(codechecker_cfg, test_config, False)
+
     ret = project.clean(test_project, test_env)
     if ret:
         sys.exit(ret)
@@ -81,19 +85,11 @@ def setup_package():
         sys.exit(1)
     print("Analyzing test project was succcessful.")
 
-    if pg_db_config:
-        print("Waiting for PotgreSQL to stop.")
-        codechecker.wait_for_postgres_shutdown(TEST_WORKSPACE)
-
     codechecker_cfg['run_names'] = [test_project_name]
 
     test_config['codechecker_cfg'] = codechecker_cfg
 
     env.export_test_cfg(TEST_WORKSPACE, test_config)
-
-    # Start the CodeChecker server.
-    print("Starting server to get results")
-    _start_server(codechecker_cfg, test_config, False)
 
 
 def teardown_package():
