+++ www/index.html
@@ -34,7 +34,7 @@
   <!-- Services -->
 
   <script type="text/javascript" src="scripts/codechecker-api/shared_types.js"></script>
-  <script type="text/javascript" src="scripts/codechecker-api/report_viewer_server_types.js"></script>
+  <script type="text/javascript" src="scripts/codechecker-api/report_server_types.js"></script>
   <script type="text/javascript" src="scripts/codechecker-api/codeCheckerDBAccess.js"></script>
 
   <!-- Configure Dojo -->
